Put `TransformOffset` `InterpolateMode` back.
