+++ inox2d/src/render.rs
@@ -2,9 +2,10 @@ use std::collections::HashMap;
 
 use glam::{vec2, Mat4, Vec2};
 
+use crate::math::transform::TransformOffset;
 use crate::mesh::Mesh;
 use crate::model::Model;
-use crate::node::data::{Composite, InoxData, MaskMode, Part, TransformOffset};
+use crate::node::data::{Composite, InoxData, MaskMode, Part};
 use crate::node::tree::InoxNodeTree;
 use crate::node::InoxNodeUuid;
 use crate::puppet::Puppet;
